Got rid of unnessessary HTTP connection between bot and API service
